test_install: move out of e2e/ alongside its fixture
The test only depends on ``minimal_install_env`` (which itself depends on
``isolated_mngr_venv`` + ``temp_host_dir`` + ``mngr_test_root_name``,
all of which already live in libs/mngr/imbue/mngr/conftest.py) and never
touches the heavy ``e2e`` fixture. Moving the test out of e2e/ requires
moving ``MinimalInstallEnv`` + ``minimal_install_env`` to the same
conftest as their dependencies, so do that as well.

The e2e/ directory is for tests that need the e2e fixture infrastructure
(remote hosts, asciinema, isolated tmux servers, Modal cleanup). These
isolated-venv tests fit better at the package root.
