+++ libs/mngr/imbue/mngr/test_install.py
@@ -10,7 +10,7 @@ import json
 
 import pytest
 
-from imbue.mngr.e2e.conftest import MinimalInstallEnv
+from imbue.mngr.conftest import MinimalInstallEnv
 
 
 @pytest.mark.release
